# Optimize assembly calls & add uint32 for amd64