--- a/dummy.s
+++ b/dummy.s
@@ -1,2 +1,0 @@
-// This empty assembly file allows to use the go:linkname directive for
-// non-assembly implementations.