--- a/endian.go
+++ b/endian.go
@@ -4,40 +4,29 @@
 // that can be found in the COPYING file.
 
 // Package endian provides helper functions to deal with network byte order.
+// The implementations reside in architecture dependent files.
 package endian
 
 // NetToHostUint16 swaps the bytes of the given value (if necessary) to turn
 // them from network byte order to host byte order.
-func NetToHostUint16(n uint16) uint16 {
-	return swapUint16(n)
-}
+func NetToHostUint16(n uint16) uint16
 
 // HostToNetUint16 swaps the bytes of the given value (if necessary) to turn
 // them from host byte order to network byte order.
-func HostToNetUint16(n uint16) uint16 {
-	return swapUint16(n)
-}
+func HostToNetUint16(n uint16) uint16
 
 // NetToHostUint32 swaps the bytes of the given value (if necessary) to turn
 // them from network byte order to host byte order.
-func NetToHostUint32(n uint32) uint32 {
-	return swapUint32(n)
-}
+func NetToHostUint32(n uint32) uint32
 
 // HostToNetUint32 swaps the bytes of the given value (if necessary) to turn
 // them from host byte order to network byte order.
-func HostToNetUint32(n uint32) uint32 {
-	return swapUint32(n)
-}
+func HostToNetUint32(n uint32) uint32
 
 // NetToHostUint64 swaps the bytes of the given value (if necessary) to turn
 // them from network byte order to host byte order.
-func NetToHostUint64(n uint64) uint64 {
-	return swapUint64(n)
-}
+func NetToHostUint64(n uint64) uint64
 
 // HostToNetUint64 swaps the bytes of the given value (if necessary) to turn
 // them from host byte order to network byte order.
-func HostToNetUint64(n uint64) uint64 {
-	return swapUint64(n)
-}
+func HostToNetUint64(n uint64) uint64