--- a/uint16_little.go
+++ b/uint16_little.go
@@ -7,6 +7,18 @@
 
 package endian
 
+import _ "unsafe"
+
+//go:linkname netToHostUint16 github.com/tsuna/endian.NetToHostUint16
+func netToHostUint16(n uint16) uint16 {
+	return swapUint16(n)
+}
+
+//go:linkname hostToNetUint16 github.com/tsuna/endian.HostToNetUint16
+func hostToNetUint16(n uint16) uint16 {
+	return swapUint16(n)
+}
+
 // swapUint16 converts a uint16 to network byte order and back.
 func swapUint16(n uint16) uint16 {
 	return (n&0x00FF)<<8 | (n&0xFF00)>>8