… uint64_amd64.go