By directly linking the assembly implementations to the `NetToHost*` and `HostToNet*` family functions, we get rid of the non-inlined double function call overhead. This makes the assembly implementations more competitive with their Go counterparts, allowing to also add the uint32 implementation for amd64.

~~The `dummy.s` file is necessary to be able to have function declarations without an implementation, and is thus required to be able to declare implementation in architecture dependent Go files linked with the `go:linkname` directive.~~

~~One downside is that we have to duplicate the code of the assembly implementations, but since they are very small and will never change, this is not a big drawback.~~

Another limitation is that we now have to import the "unsafe" package, required to be able to use the `go:linkname` directive. This could lead to functions being mistakenly defined without implementation, but no new function will be added to this library, and new potential assembly implementations will not be affected.

Here are the benchmark results:

Before
```console
$ go test . -bench=.
goos: linux
goarch: amd64
pkg: github.com/tsuna/endian
cpu: 13th Gen Intel(R) Core(TM) i9-13900
Benchmark16_1000-32    	 2421727	       493.7 ns/op
Benchmark32_1000-32    	 1643253	       728.9 ns/op
Benchmark64_1000-32    	 1412155	       844.7 ns/op
PASS
ok  	github.com/tsuna/endian	3.590s
```

After
```console
$ go test . -bench=.
goos: linux
goarch: amd64
pkg: github.com/tsuna/endian
cpu: 13th Gen Intel(R) Core(TM) i9-13900
Benchmark16_1000-32    	 2387834	       496.4 ns/op
Benchmark32_1000-32    	 2446610	       490.2 ns/op
Benchmark64_1000-32    	 2395137	       495.9 ns/op
PASS
ok  	github.com/tsuna/endian	3.576s
```

The noasm version performances have not changed:
```console
$ go test . -bench=. -tags noasm
goos: linux
goarch: amd64
pkg: github.com/tsuna/endian
cpu: 13th Gen Intel(R) Core(TM) i9-13900
Benchmark16_1000-32    	 2344089	       499.6 ns/op
Benchmark32_1000-32    	 1627462	       736.2 ns/op
Benchmark64_1000-32    	  948756	      1269 ns/op
PASS
ok  	github.com/tsuna/endian	3.577s
```