@n-peugnet 2025-09-08
Hi @tsuna, here is just a quick ping about this PR. Did you see my previous comment? And did you have time to think about your preferred way to do this optimization? Or if you want to optimize it at all?

I also have an arm64 version ready and was waiting to see which way you would choose before making the pull request.